Pairwise Comparison of Clusterings

Compare a set of clusterings by calculating their pairwise average element-centric clustering similarities.

element_sim_matrix(clustering_list, output_type = "matrix")

Arguments

clustering_list
A list of Clustering objects to be compared with element-centric similarity.

output_type
A string specifying whether the output should be a matrix or a data.frame.

Value

A matrix or data.frame containing the pairwise ECS values.

References

Gates, A. J., Wood, I. B., Hetrick, W. P., & Ahn, Y. Y. (2019). Element-centric clustering comparison unifies overlaps and hierarchy. Scientific reports, 9(1), 1-13. https://doi.org/10.1038/s41598-019-44892-y

Examples

clustering.list = list()
for (i in 1:20){
  km.res = kmeans(mtcars, 3)$cluster
  clustering.list[[i]] = create_clustering(km.res)
}
element_sim_matrix(clustering.list, output_type='matrix')

#>       [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,]    1 0.5396205 1.0000000 0.5503724 0.5396205 1.0000000 0.5396205
#>  [2,]   NA 1.0000000 0.5396205 0.7012775 1.0000000 0.5396205 1.0000000
#>  [3,]   NA        NA 1.0000000 0.5503724 0.5396205 1.0000000 0.5396205
#>  [4,]   NA        NA        NA 1.0000000 0.7012775 0.5503724 0.7012775
#>  [5,]   NA        NA        NA        NA 1.0000000 0.5396205 1.0000000
#>  [6,]   NA        NA        NA        NA        NA 1.0000000 0.5396205
#>  [7,]   NA        NA        NA        NA        NA        NA 1.0000000
#>  [8,]   NA        NA        NA        NA        NA        NA        NA
#>  [9,]   NA        NA        NA        NA        NA        NA        NA
#> [10,]   NA        NA        NA        NA        NA        NA        NA
#> [11,]   NA        NA        NA        NA        NA        NA        NA
#> [12,]   NA        NA        NA        NA        NA        NA        NA
#> [13,]   NA        NA        NA        NA        NA        NA        NA
#> [14,]   NA        NA        NA        NA        NA        NA        NA
#> [15,]   NA        NA        NA        NA        NA        NA        NA
#> [16,]   NA        NA        NA        NA        NA        NA        NA
#> [17,]   NA        NA        NA        NA        NA        NA        NA
#> [18,]   NA        NA        NA        NA        NA        NA        NA
#> [19,]   NA        NA        NA        NA        NA        NA        NA
#> [20,]   NA        NA        NA        NA        NA        NA        NA
#>            [,8]      [,9]     [,10]     [,11]     [,12]     [,13]     [,14]
#>  [1,] 0.5503724 1.0000000 0.5396205 0.5396205 1.0000000 0.5396205 0.5396205
#>  [2,] 0.7012775 0.5396205 1.0000000 1.0000000 0.5396205 1.0000000 1.0000000
#>  [3,] 0.5503724 1.0000000 0.5396205 0.5396205 1.0000000 0.5396205 0.5396205
#>  [4,] 1.0000000 0.5503724 0.7012775 0.7012775 0.5503724 0.7012775 0.7012775
#>  [5,] 0.7012775 0.5396205 1.0000000 1.0000000 0.5396205 1.0000000 1.0000000
#>  [6,] 0.5503724 1.0000000 0.5396205 0.5396205 1.0000000 0.5396205 0.5396205
#>  [7,] 0.7012775 0.5396205 1.0000000 1.0000000 0.5396205 1.0000000 1.0000000
#>  [8,] 1.0000000 0.5503724 0.7012775 0.7012775 0.5503724 0.7012775 0.7012775
#>  [9,]        NA 1.0000000 0.5396205 0.5396205 1.0000000 0.5396205 0.5396205
#> [10,]        NA        NA 1.0000000 1.0000000 0.5396205 1.0000000 1.0000000
#> [11,]        NA        NA        NA 1.0000000 0.5396205 1.0000000 1.0000000
#> [12,]        NA        NA        NA        NA 1.0000000 0.5396205 0.5396205
#> [13,]        NA        NA        NA        NA        NA 1.0000000 1.0000000
#> [14,]        NA        NA        NA        NA        NA        NA 1.0000000
#> [15,]        NA        NA        NA        NA        NA        NA        NA
#> [16,]        NA        NA        NA        NA        NA        NA        NA
#> [17,]        NA        NA        NA        NA        NA        NA        NA
#> [18,]        NA        NA        NA        NA        NA        NA        NA
#> [19,]        NA        NA        NA        NA        NA        NA        NA
#> [20,]        NA        NA        NA        NA        NA        NA        NA
#>           [,15]     [,16]     [,17]     [,18]     [,19]     [,20]
#>  [1,] 0.5396205 0.5396205 1.0000000 0.5503724 1.0000000 0.5396205
#>  [2,] 1.0000000 1.0000000 0.5396205 0.7012775 0.5396205 1.0000000
#>  [3,] 0.5396205 0.5396205 1.0000000 0.5503724 1.0000000 0.5396205
#>  [4,] 0.7012775 0.7012775 0.5503724 1.0000000 0.5503724 0.7012775
#>  [5,] 1.0000000 1.0000000 0.5396205 0.7012775 0.5396205 1.0000000
#>  [6,] 0.5396205 0.5396205 1.0000000 0.5503724 1.0000000 0.5396205
#>  [7,] 1.0000000 1.0000000 0.5396205 0.7012775 0.5396205 1.0000000
#>  [8,] 0.7012775 0.7012775 0.5503724 1.0000000 0.5503724 0.7012775
#>  [9,] 0.5396205 0.5396205 1.0000000 0.5503724 1.0000000 0.5396205
#> [10,] 1.0000000 1.0000000 0.5396205 0.7012775 0.5396205 1.0000000
#> [11,] 1.0000000 1.0000000 0.5396205 0.7012775 0.5396205 1.0000000
#> [12,] 0.5396205 0.5396205 1.0000000 0.5503724 1.0000000 0.5396205
#> [13,] 1.0000000 1.0000000 0.5396205 0.7012775 0.5396205 1.0000000
#> [14,] 1.0000000 1.0000000 0.5396205 0.7012775 0.5396205 1.0000000
#> [15,] 1.0000000 1.0000000 0.5396205 0.7012775 0.5396205 1.0000000
#> [16,]        NA 1.0000000 0.5396205 0.7012775 0.5396205 1.0000000
#> [17,]        NA        NA 1.0000000 0.5503724 1.0000000 0.5396205
#> [18,]        NA        NA        NA 1.0000000 0.5503724 0.7012775
#> [19,]        NA        NA        NA        NA 1.0000000 0.5396205
#> [20,]        NA        NA        NA        NA        NA 1.0000000