Graph building Parameters Importance

Creates an object that contain stability measurements and assessments when changing the values of different parameters involved in the Graph Building step, namely the base embedding, the graph type and the number of neighbours.

get_nn_importance(
  object,
  n_neigh_sequence,
  n_repetitions = 30,
  seed_sequence = NULL,
  graph_reduction_type = "PCA",
  ecs_thresh = 1,
  ncores = 1,
  transpose = (graph_reduction_type == "PCA"),
  graph_type = 2,
  algorithm = 4,
  ...
)

Arguments

object

A matrix object. If the graph reduction type is PCA, the object should be an expression matrix, with features on rows and observations on columns; in the case of UMAP, the user could also provide a matrix associated to a PCA embedding.

n_neigh_sequence

A sequence of the number of nearest neighbours.

n_repetitions

The number of repetitions of applying the pipeline with different seeds; ignored if seed_sequence is provided by the user.

seed_sequence

A custom seed sequence; if the value is NULL, the sequence will be built starting from 1 with a step of 100.

graph_reduction_type

The graph reduction type, denoting if the graph should be built on either the PCA or the UMAP embedding.

ecs_thresh

The ECS threshold used for merging similar clusterings.

ncores

The number of parallel R instances that will run the code. If the value is set to 1, the code will be run sequentially.

transpose

Decides whether the input object will be transposed or not.

graph_type

Argument indicating whether the graph should be unweighted (0), weighted (1) or both (2).

algorithm

An index indicating which community detection algorithm will be used: Louvain (1), Louvain refined (2), SLM (3) or Leiden (4). More details can be found in the Seurat's FindClusters function.

...

Additional arguments passed to the irlba::irlba or the uwot::umap method, depending on the value of graph_reduction_type.

Value

A list having three fields:

n_neigh_k_corresp - list containing the number of the clusters obtained by running the pipeline multiple times with different seed, number of neighbors and graph type (weighted vs unweigted)

n_neigh_ec_consistency - list containing the EC consistency of the partitions obtained at multiple runs when changing the number of neighbors or the graph type

n_different_partitions - the number of different partitions obtained by each number of neighbors

Examples

#' # get the PCA embedding of the data
pca_embedding = irlba::irlba(as.matrix(mtcars), nv = 2)
pca_embedding = pca_embedding$u %*% diag(pca_embedding$d)
rownames(pca_embedding) = rownames(mtcars)

nn_importance_obj = get_nn_importance(object = pca_embedding,
    n_neigh_sequence = c(5,15),
    n_repetitions = 1,
    graph_reduction_type = "UMAP",
    algorithm = 1,
    # the following parameters are used by the umap function and are not mandatory
    min_dist = 0.3,
    n_neighbors = 5,
    metric = "cosine",
    approx_pow = TRUE,
    n_epochs = 0,
    init = "random")