UI - Dimensionality reduction module

Creates the UI interface for the dimensionality reduction module inside the ClustAssess Shiny application.

Usage

ui_dimensionality_reduction(id)

Arguments

id

The id of the module, used to identify the UI elements.

Note

This function should not be called directly, but in the context of the app that is created using the write_shiny_app function.