The Element-Centric Clustering Similarity for each Element

Calculates the element-wise element-centric similarity between two clustering results.

element_sim_elscore(
  clustering1,
  clustering2,
  alpha = 0.9,
  r_cl1 = 1,
  rescale_path_type_cl1 = "max",
  ppr_implementation_cl1 = "prpack",
  dist_rescaled_cl1 = FALSE,
  row_normalize_cl1 = TRUE,
  r_cl2 = 1,
  rescale_path_type_cl2 = "max",
  ppr_implementation_cl2 = "prpack",
  dist_rescaled_cl2 = FALSE,
  row_normalize_cl2 = TRUE
)

Arguments

clustering1

The first clustering result, which can be one of:

A numeric/character/factor vector of cluster labels for each element.

A samples x clusters matrix/Matrix::Matrix of nonzero membership values.

An hclust object.

clustering2

The second clustering result, which can be one of:

A numeric/character/factor vector of cluster labels for each element.

A samples x clusters matrix/Matrix::Matrix of nonzero membership values.

An hclust object.

alpha

A numeric giving the personalized PageRank damping factor; 1 - alpha is the restart probability for the PPR random walk.

r_cl1

A numeric hierarchical scaling parameter for the first clustering.

rescale_path_type_cl1

A string; rescale the hierarchical height of the first clustering by:

"max" : the maximum path from the root.

"min" : the minimum path form the root.

"linkage" : use the linkage distances in the clustering.

ppr_implementation_cl1

Choose a implementation for personalized page-rank calculation for the first clustering:

"prpack": use PPR algorithms in igraph.

"power_iteration": use power_iteration method.

dist_rescaled_cl1

A logical: if TRUE, the linkage distances of the first clustering are linearly rescaled to be in-between 0 and 1.

row_normalize_cl1

Whether to normalize all rows in the first clustering so they sum to one before calculating ECS. It is recommended to set this to TRUE, which will lead to slightly different ECS values compared to clusim.

r_cl2

A numeric hierarchical scaling parameter for the second clustering.

rescale_path_type_cl2

A string; rescale the hierarchical height of the second clustering by:

"max" : the maximum path from the root.

"min" : the minimum path form the root.

"linkage" : use the linkage distances in the clustering.

ppr_implementation_cl2

Choose a implementation for personalized page-rank calculation for the second clustering:

"prpack": use PPR algorithms in igraph.

"power_iteration": use power_iteration method.

dist_rescaled_cl2

A logical: if TRUE, the linkage distances of the second clustering are linearly rescaled to be in-between 0 and 1.

row_normalize_cl2

Whether to normalize all rows in the second clustering so they sum to one before calculating ECS. It is recommended to set this to TRUE, which will lead to slightly different ECS values compared to clusim.

Value

Vector of element-centric similarity between the two clusterings for each element.

References

Gates, A. J., Wood, I. B., Hetrick, W. P., & Ahn, Y. Y. (2019). Element-centric clustering comparison unifies overlaps and hierarchy. Scientific reports, 9(1), 1-13. https://doi.org/10.1038/s41598-019-44892-y

Examples

km.res = kmeans(iris[,1:4], centers=8)$cluster
hc.res = hclust(dist(iris[,1:4]))
element_sim_elscore(km.res, hc.res)
#>         1         2         3         4         5         6         7         8 
#> 0.3878285 0.3589140 0.3838371 0.3816812 0.3657773 0.1731130 0.3645628 0.3878285 
#>         9        10        11        12        13        14        15        16 
#> 0.3670503 0.3589140 0.1543667 0.4020187 0.3558931 0.3644576 0.2178441 0.2178441 
#>        17        18        19        20        21        22        23        24 
#> 0.2185560 0.3878285 0.1731130 0.3970228 0.3748286 0.3970228 0.3271692 0.4011195 
#>        25        26        27        28        29        30        31        32 
#> 0.4020187 0.3558931 0.4011195 0.3783396 0.3783396 0.3386682 0.3386682 0.3748286 
#>        33        34        35        36        37        38        39        40 
#> 0.2390505 0.2390505 0.3589140 0.2885336 0.3709280 0.3657773 0.3670503 0.3878285 
#>        41        42        43        44        45        46        47        48 
#> 0.3907523 0.2739563 0.3779919 0.4076047 0.3982183 0.3589140 0.4055177 0.3816812 
#>        49        50        51        52        53        54        55        56 
#> 0.3596746 0.3907523 0.3286844 0.2873330 0.3286844 0.4764463 0.3290936 0.1716152 
#>        57        58        59        60        61        62        63        64 
#> 0.2873330 0.3677578 0.3290936 0.4745789 0.3742716 0.1535189 0.4435410 0.3292467 
#>        65        66        67        68        69        70        71        72 
#> 0.4632903 0.3171300 0.1819448 0.4696937 0.3033067 0.4750907 0.3994726 0.4375253 
#>        73        74        75        76        77        78        79        80 
#> 0.3989538 0.3359173 0.2769332 0.3171300 0.3172690 0.3025440 0.3337228 0.4632903 
#>        81        82        83        84        85        86        87        88 
#> 0.4690584 0.4690584 0.4640232 0.3925891 0.1819448 0.3090685 0.3168525 0.2456608 
#>        89        90        91        92        93        94        95        96 
#> 0.4752405 0.4764463 0.4074367 0.3292467 0.4640232 0.3677578 0.4717717 0.4652972 
#>        97        98        99       100       101       102       103       104 
#> 0.4652972 0.2769332 0.3764827 0.4717717 0.3817927 0.3932686 0.3548650 0.4213612 
#>       105       106       107       108       109       110       111       112 
#> 0.4213612 0.3689191 0.3829617 0.3469864 0.3966005 0.3755302 0.4050782 0.2042484 
#>       113       114       115       116       117       118       119       120 
#> 0.4118917 0.3981969 0.3963147 0.4078736 0.4141239 0.3755302 0.3585443 0.3150768 
#>       121       122       123       124       125       126       127       128 
#> 0.3887176 0.4071223 0.3689191 0.3977770 0.3981992 0.3605472 0.3977770 0.3897570 
#>       129       130       131       132       133       134       135       136 
#> 0.4141239 0.3605472 0.3469864 0.3755302 0.4141239 0.3925891 0.3733406 0.3755302 
#>       137       138       139       140       141       142       143       144 
#> 0.3788777 0.4141239 0.3897570 0.4118917 0.3913903 0.4118917 0.3932686 0.3887176 
#>       145       146       147       148       149       150 
#> 0.3913903 0.4118917 0.3770036 0.4050782 0.3788777 0.3945777