The Element-Centric Clustering Similarity for each Element

Calculates the element-wise element-centric similarity between two clustering results.

element_sim_elscore(
  clustering1,
  clustering2,
  alpha = 0.9,
  r_cl1 = 1,
  rescale_path_type_cl1 = "max",
  ppr_implementation_cl1 = "prpack",
  dist_rescaled_cl1 = FALSE,
  row_normalize_cl1 = TRUE,
  r_cl2 = 1,
  rescale_path_type_cl2 = "max",
  ppr_implementation_cl2 = "prpack",
  dist_rescaled_cl2 = FALSE,
  row_normalize_cl2 = TRUE
)

Arguments

clustering1

The first clustering result, which can be one of:

A numeric/character/factor vector of cluster labels for each element.

A samples x clusters matrix/Matrix::Matrix of nonzero membership values.

An hclust object.

clustering2

The second clustering result, which can be one of:

A numeric/character/factor vector of cluster labels for each element.

A samples x clusters matrix/Matrix::Matrix of nonzero membership values.

An hclust object.

alpha

A numeric giving the personalized PageRank damping factor; 1 - alpha is the restart probability for the PPR random walk.

r_cl1

A numeric hierarchical scaling parameter for the first clustering.

rescale_path_type_cl1

A string; rescale the hierarchical height of the first clustering by:

"max" : the maximum path from the root.

"min" : the minimum path form the root.

"linkage" : use the linkage distances in the clustering.

ppr_implementation_cl1

Choose a implementation for personalized page-rank calculation for the first clustering:

"prpack": use PPR algorithms in igraph.

"power_iteration": use power_iteration method.

dist_rescaled_cl1

A logical: if TRUE, the linkage distances of the first clustering are linearly rescaled to be in-between 0 and 1.

row_normalize_cl1

Whether to normalize all rows in the first clustering so they sum to one before calculating ECS. It is recommended to set this to TRUE, which will lead to slightly different ECS values compared to clusim.

r_cl2

A numeric hierarchical scaling parameter for the second clustering.

rescale_path_type_cl2

A string; rescale the hierarchical height of the second clustering by:

"max" : the maximum path from the root.

"min" : the minimum path form the root.

"linkage" : use the linkage distances in the clustering.

ppr_implementation_cl2

Choose a implementation for personalized page-rank calculation for the second clustering:

"prpack": use PPR algorithms in igraph.

"power_iteration": use power_iteration method.

dist_rescaled_cl2

A logical: if TRUE, the linkage distances of the second clustering are linearly rescaled to be in-between 0 and 1.

row_normalize_cl2

Whether to normalize all rows in the second clustering so they sum to one before calculating ECS. It is recommended to set this to TRUE, which will lead to slightly different ECS values compared to clusim.

Value

Vector of element-centric similarity between the two clusterings for each element.

References

Gates, A. J., Wood, I. B., Hetrick, W. P., & Ahn, Y. Y. (2019). Element-centric clustering comparison unifies overlaps and hierarchy. Scientific reports, 9(1), 1-13. https://doi.org/10.1038/s41598-019-44892-y

Examples

km.res = kmeans(iris[,1:4], centers=8)$cluster
hc.res = hclust(dist(iris[,1:4]))
element_sim_elscore(km.res, hc.res)
#>          1          2          3          4          5          6          7 
#> 0.51062339 0.50954967 0.49531045 0.48938539 0.51748731 0.33067968 0.51282743 
#>          8          9         10         11         12         13         14 
#> 0.51062339 0.49919403 0.50954967 0.35238739 0.37377785 0.51717729 0.51034517 
#>         15         16         17         18         19         20         21 
#> 0.31377671 0.31377671 0.32254634 0.51062339 0.33067968 0.32848901 0.34610343 
#>         22         23         24         25         26         27         28 
#> 0.32848901 0.49148473 0.36627708 0.37377785 0.51717729 0.36627708 0.16043044 
#>         29         30         31         32         33         34         35 
#> 0.51505357 0.50914401 0.50914401 0.34610343 0.32590888 0.32590888 0.50954967 
#>         36         37         38         39         40         41         42 
#> 0.50838795 0.35038499 0.51748731 0.49919403 0.51062339 0.51574284 0.42681569 
#>         43         44         45         46         47         48         49 
#> 0.50218595 0.37169513 0.31023693 0.50954967 0.32629990 0.48938539 0.35238739 
#>         50         51         52         53         54         55         56 
#> 0.51574284 0.39712276 0.38883418 0.39712276 0.51263990 0.39901884 0.50763058 
#>         57         58         59         60         61         62         63 
#> 0.38883418 0.39088001 0.39901884 0.52643208 0.41153617 0.51868674 0.49906433 
#>         64         65         66         67         68         69         70 
#> 0.40535658 0.51409810 0.39428345 0.50763058 0.53340573 0.24653654 0.51294180 
#>         71         72         73         74         75         76         77 
#> 0.36587735 0.51868674 0.38077910 0.39660042 0.39462504 0.39428345 0.39719720 
#>         78         79         80         81         82         83         84 
#> 0.38419606 0.40632445 0.51409810 0.50674233 0.50674233 0.52168669 0.38040131 
#>         85         86         87         88         89         90         91 
#> 0.50763058 0.38346879 0.39682948 0.30056420 0.52689686 0.51263990 0.50763058 
#>         92         93         94         95         96         97         98 
#> 0.40535658 0.52168669 0.39088001 0.52938883 0.52041953 0.52041953 0.39462504 
#>         99        100        101        102        103        104        105 
#> 0.40000525 0.52938883 0.36929795 0.37561418 0.24538234 0.40230846 0.40230846 
#>        106        107        108        109        110        111        112 
#> 0.30400914 0.43395801 0.24435746 0.07363641 0.30618792 0.39382441 0.36444712 
#>        113        114        115        116        117        118        119 
#> 0.39772697 0.37675169 0.35397987 0.39284928 0.39965860 0.30618792 0.28224743 
#>        120        121        122        123        124        125        126 
#> 0.25626715 0.38120351 0.36961918 0.30400914 0.37599274 0.38618677 0.26826176 
#>        127        128        129        130        131        132        133 
#> 0.37599274 0.36644908 0.39965860 0.26826176 0.24435746 0.30618792 0.39965860 
#>        134        135        136        137        138        139        140 
#> 0.38040131 0.33457327 0.30618792 0.37096474 0.39965860 0.36644908 0.39772697 
#>        141        142        143        144        145        146        147 
#> 0.37937007 0.39772697 0.37561418 0.38120351 0.37937007 0.39772697 0.36444712 
#>        148        149        150 
#> 0.39382441 0.37096474 0.35387715